Fix lock contention when posting tasks.
Before this patch, multiple threads posting to some other thread would
experience lock contention solely for the 'notify' step.  Notify is only
necessary when the thread is 'blocked' and idle, otherwise it is a
waste.

With these changes, a lock has been replaced with an atomic read of a
variable that is only true when the other thread is 'blocking'.  This
should scale much better.